requirements.txt
# Core LLM
openai==1.40.0
langchain==0.2.12
langchain-openai==0.1.20
langchain-community==0.2.11
langgraph==0.2.0

# Email
google-api-python-client==2.137.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# CV Parsing
PyPDF2==3.0.1
python-docx==1.1.0

# NLP
spacy==3.7.5
chromadb==0.5.5
sentence-transformers==2.7.0

# Web Automation
selenium==4.23.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
requests==2.32.3

# Database
sqlalchemy==2.0.31

# Scheduler
apscheduler==3.10.4

# Dashboard
streamlit==1.37.0

# Notifications
python-telegram-bot==21.4

# Utils
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
rich==13.7.1
tiktoken==0.7.0
